requirements.txt
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0
qdrant-client==1.7.0
openai>=1.0.0
# google-generativeai removed - using REST API via httpx instead
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.0.0.post2
sqlalchemy==2.0.23
typing-extensions>=4.8.0
nest-asyncio>=1.5.8